reorganized test_views into three files
